Added code in makefile to handle size of mOS.bin binary.
On QEMU, the new error handling code in the boot sector will fail since the file
generated is not 512*42 bytes large (some calls to the INT 0x13/AH=0x02
interrrupt will fail, due to the size of the drive created/mounted by
QEMU).

The Makefile was updated both to make the mOS.bin file large enough if
necessary (using truncate), and to give an error if the generated mOS.bin file
is too large (larger than 512*42 bytes).

Note: This has not been an issue though for booting off of floppy disks on real
hardware.